# packages/core/api.yaml
openapi: 3.0.0
paths:
...
  /v1/telco/check-coverage:
    post:
      operationId: telcoCheckCoverageV1
      parameters:
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckCoverageDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicCoverageDto'
        '401':
          description: Unauthorized
        '429':
          description: Rate limited
      security:
        - bearer: []
        - cookie: []
      tags: &ref_1
        - Telco
  /v1/telco/check-portin:
    post:
      operationId: telcoCheckPortInV1
      parameters:
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckPortInDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPortInDto'
        '401':
          description: Unauthorized
        '429':
          description: Rate limited
      security:
        - bearer: []
        - cookie: []
      tags: *ref_1
...
  /v1/auth/open-id:
    get:
      operationId: openIdFindStrategiesV1
      parameters:
        - name: admin
          required: true
          in: query
          schema:
            type: boolean
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuthStrategyDto'
        '429':
          description: Rate limited
      tags: &ref_3
        - Auth
  /v1/auth/open-id/{strategy}/login:
    post:
      operationId: openIdLoginRedeemV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthOpenIdLoginRedeemDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_3
  /v1/auth/open-id/{strategy}/exchange:
    get:
      operationId: openIdExchangeStartV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicExchangeDto'
        '429':
          description: Rate limited
      tags: *ref_3
    post:
      operationId: openIdExchangeCallbackV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthOpenIdExchangeCallbackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_3
...
info:
  title: Fliggs
  description: This is our cool new API
  version: '0.1'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PublicSessionTokensDto:
      type: object
      properties:
        token:
          type: string
        expires:
          type: string
        refreshToken:
          type: string
        refreshExpires:
          type: string
      required:
        - token
        - expires
        - refreshToken
        - refreshExpires
...
    CheckCoverageDto:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        line3:
          type: string
        line4:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
        country:
          type: string
      required:
        - line1
        - line2
        - line3
        - line4
        - city
        - postalCode
        - province
        - country
    PublicCoverageDto:
      type: object
      properties:
        coverage:
          type: number
          nullable: true
      required:
        - coverage
    CheckPortInDto:
      type: object
      properties:
        msisdn:
          type: string
      required:
        - msisdn
    PublicPortInDto:
      type: object
      properties:
        isEligible:
          type: boolean
      required:
        - isEligible
...
    AuthProvider:
      type: string
      enum:
        - local
        - google
        - apple
        - open-id
    PublicAuthStrategyDto:
      type: object
      properties:
        provider:
          allOf:
            - $ref: '#/components/schemas/AuthProvider'
        name:
          type: string
        title:
          type: string
        tags:
          nullable: true
          type: array
          items:
            type: string
      required:
        - provider
        - name
        - title
...
    AuthOpenIdLoginRedeemDto:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    PublicExchangeDto:
      type: object
      properties:
        issuerUrl:
          type: string
        clientId:
          type: string
        redirectUrl:
          type: string
        scope:
          type: string
      required:
        - issuerUrl
        - clientId
        - redirectUrl
        - scope
    AuthOpenIdExchangeCallbackDto:
      type: object
      properties:
        idToken:
          type: string
      required:
        - idToken
...